Initial Commit
This commit includes all of the files necessary to use Django 2.1 with
React on Heroku, set up to serve a multi-page application. It uses
Waitress as the WSGI server and supplies a working configuration of
Celery suitable for a single worker using gevent as the execution pool
that also runs the beat process for cron-like scheduling of jobs.
This should be sufficient for most small proof-of-concept or demo
applications. Productionization may require further configuration of
Celery, Waitress, and Django.